refactor(chat): give provider message conversion its own module
Move convertUiMessagesToProviderModelMessages and its private helpers
(buildToolNameMap, resolveRawToolResultPart, shouldSkipTransientToolCall,
convertSystemMessage, convertUserMessage, convertAssistantMessage,
convertToolMessage, ProviderToolResultContent) out of conversation.ts
into src/chat/provider-message-conversion.ts, verbatim.

conversation.ts keeps the persisted-conversation format and how UI
messages become stored parts; provider-message-conversion.ts is now the
single owner of turning replay history into the ordered message list a
provider sees. Consumers (message-prep.ts, compat.ts,
chat-request.test.ts, conversation.test.ts) repoint directly at the new
module rather than via a re-export from conversation.ts.
