+++ src/chat/message-prep.ts
@@ -1,11 +1,11 @@
 import {
-  convertUiMessagesToProviderModelMessages,
   copyProviderModelMessageSourceId,
   getStringField,
   isReasoningPart,
   isToolCallPart,
   isToolResultPart,
 } from "./conversation.ts";
+import { convertUiMessagesToProviderModelMessages } from "./provider-message-conversion.ts";
 import { findProviderVisibleToolReplayMatches } from "./tool-replay-reconciliation.ts";
 import {
   buildDataFileAnnotation,
